+++ include/xsk/gsc/compiler.hpp
@@ -23,8 +23,8 @@ class compiler
     std::vector<scope*> continue_blks_;
     std::string animname_;
     sourcepos debug_pos_;
-    u32 index_;
-    u32 label_idx_;
+    usize index_;
+    usize label_idx_;
     bool can_break_;
     bool can_continue_;
     bool developer_thread_;
